Cleanup documentation and style
