+++ .gitignore
@@ -3,3 +3,4 @@ dist/
 yarn-error.log
 docs/modules/*.ts.md
 docs/_site
+docs/
